Added minor functionality and fixed 2 minor errors
Updated the Wrappers to ensure PREP folders are found (even if the latest version of EEGLAB is used, which seems to cause an issue).

Fixed a dimension selection issue when determining if PREP rejected too many electrodes in the metrics section of the Wrappers (thank you to Tyson Perez for noting this).

Added an option in RELAX_beta to downsample the data prior to cleaning (but after low pass filtering).

Fixed an issue with blink amplitude ratio metric, where it was being calculated as if the sampling rate was always 1000Hz. Now BAR is calculated based on the file specific sample rate.